improve the project's organizational structure
